move tests over to mocha
