+++ core/.mocharc.json
@@ -1,5 +1,6 @@
 {
   "extension": ["ts"],
   "spec": ["tests/**/*.spec.ts"],
-  "require": ["ts-node/register"]
+  "require": ["ts-node/register"],
+  "timeout": 35000
 }
